WEBAPP-121: Move CategoriesTable into TitledCategoriesTable
