--- src/components/Content/CategoriesTable.css
@@ -1,28 +1,0 @@
-@import url('../Helper/Helper.css');
-
-.thumbnail {
-  display: block;
-  width: 150px;
-  max-width: 100%;
-  max-height: 100%;
-  margin: 0 auto;
-  transition: transform 0.2s;
-}
-
-.thumbnail:hover {
-  transform: scale(1.01);
-}
-
-.caption {
-  margin-top: 5px;
-  text-align: center;
-}
-
-.category {
-  padding-top: 10px;
-  padding-bottom: 10px;
-}
-
-.category a {
-  @mixin remove-a;
-}
