+++ src/components/Content/TitledCategoriesTable.css
@@ -1,0 +1,28 @@
+@import url('../Helper/Helper.css');
+
+.category-thumbnail {
+  display: block;
+  width: 150px;
+  max-width: 100%;
+  max-height: 100%;
+  margin: 0 auto;
+  transition: transform 0.2s;
+}
+
+.category-thumbnail:hover {
+  transform: scale(1.01);
+}
+
+.category-title {
+  margin-top: 5px;
+  text-align: center;
+}
+
+.category {
+  padding-top: 10px;
+  padding-bottom: 10px;
+}
+
+.category a {
+  @mixin remove-a;
+}
