+++ src/components/Content/TitledCategoriesTable.js
@@ -1,11 +1,32 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import { Col, Row } from 'react-flexbox-grid'
+import { Link } from 'redux-little-router'
 
 import PageModel from 'endpoints/models/PageModel'
-import CategoriesTable from './CategoriesTable'
 import LOCATIONS_ENDPOINT from 'endpoints/location'
 import withFetcher from 'endpoints/withFetcher'
 import Caption from './Caption'
+
+import style from './TitledCategoriesTable.css'
+
+class Category extends React.Component {
+  static propTypes = {
+    page: PropTypes.instanceOf(PageModel).isRequired,
+    url: PropTypes.string.isRequired
+  }
+
+  render () {
+    return (
+      <Col xs={6} sm={4} className={style.category}>
+        <Link href={this.props.url}>
+          <img className={style.categoryThumbnail} src={this.props.page.thumbnail}/>
+          <div className={style.categoryTitle}>{this.props.page.title}</div>
+        </Link>
+      </Col>
+    )
+  }
+}
 
 class TitledCategoriesTable extends React.Component {
   static propTypes = {
@@ -24,7 +45,9 @@ class TitledCategoriesTable extends React.Component {
     return (
       <div>
         <Caption title={this.getTitle()}/>
-        <CategoriesTable pages={this.props.pages}/>
+        <Row>
+          {this.props.pages.map(({page, url}) => <Category key={page.id} url={url} page={page}/>)}
+        </Row>
       </div>
     )
   }
